Refactor AI module: rename imports for consistency and implement base AI class; add GroqAI and OllamaAI generators with YAML prompt loading
